+++ src/Tokenizers/tree_tokenizer.py
@@ -40,7 +40,7 @@ class Tree_Tokenizer(Base_Tokenizer):
                 else:
                     lis.append('<unu> '+comment['body'])
             
-        token_ids = np.array( self.get_token_ids(self.batch_encode_plus(lis)), dtype=np.int16)
+        token_ids = np.array( self.get_token_ids(self.batch_encode_plus(lis)), dtype=np.int32)
         
         i=0
         tree['tokenized_inputs'] = token_ids[i]
